fix: move unet model to unet folder
